crypto: zstd: support negative compression level
zstd supports negative compression level, which will further reduce
the compression ratio and increase the compression and decompression
speed.

But please note that according to [zstd official benchmark results](https://github.com/facebook/zstd#benchmarks),
when you set the compression level to "-3", the compression ratio and
compression speed are very close to lz4, and the decompression speed
is only about 50% of lz4.

Therefore, it is best to use lz4 at this time, and it also tells us
not to blindly set too low a compression level for zstd.

Signed-off-by: fjrXTR <fajarslebew31@gmail.com>